css/css-images/parsing/object-fit-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Images Module Level 3 Test: getComputedValue().objectFit</title>
<link rel="help" href="https://drafts.csswg.org/css-images-3/#propdef-object-fit">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("object-fit", "contain");
test_computed_value("object-fit", "cover");
test_computed_value("object-fit", "fill");
test_computed_value("object-fit", "none");
test_computed_value("object-fit", "scale-down");
</script>
</body>
</html>
